pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Parham Mohammad Panahi", email = "panahiparham@gmail.com"},
]
dependencies = [
    "numpy>=1.24.3",
    "pillow>=9.5.0",
    "opencv-python>=4.7.0.72",
]
requires-python = ">=3.10"
license = {text = "y"}
